pyproject.toml
[tool.poetry]
name = "djangoauth"
version = "0.1.0"
description = ""
authors = ["ratchaphon1412 <ratchaphon.h111@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
markdown = "^3.5.1"
django-filter = "^23.5"
dj-database-url = "^2.1.0"
django-environ = "^0.11.2"
requests = "^2.31.0"
hvac = "^2.0.0"
mysqlclient = "^2.2.1"
djangorestframework-simplejwt = "^5.3.1"
django-role-permissions = "^3.2.0"
python-gettext = "^5.0"
django-postgres-vault = "^0.1.2"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
